+++ test/sails_forth/clojurify_test.clj
@@ -30,8 +30,34 @@
                                :scale 2
                               :precision 18}
                              500M))))
+  (testing "percent"
+    (is (= 0.0618M (parse-value {:type "percent"
+                                 :scale 2
+                                 :precision 4}
+                                6.18M))))
   (testing "int"
     (is (= 500 (parse-value {:type "int"} 500M))))
+  (testing "other"
+    (is (= "foo" (parse-value {} "foo")))))
+
+(deftest test-render-value
+  (testing "datetime"
+    (is (= (render-value {:type "datetime"}
+                         (time/date-time 2015 1 1 12 30 15 500))
+           "2015-01-01T12:30:15.500Z"))
+    (is (= (render-value {:type "datetime"}
+                         "2015-01-01T12:30:15.500Z")
+           "2015-01-01T12:30:15.500Z")))
+  (testing "date"
+    (is (= (render-value {:type "date"}
+                         (time/local-date 2015 2 1))
+           "2015-02-01")))
+  (testing "percent"
+    (is (= (render-value {:type "percent"
+                          :scale 2
+                          :precision 4}
+                         0.0618M)
+           6.18M)))
   (testing "other"
     (is (= "foo" (parse-value {} "foo")))))